pyproject.toml
[project]
name = "db-magic"
version = "0.1.0"
description = "Jupyter magic commands for Databricks SQL queries"
requires-python = ">=3.9"
dependencies = [
    "databricks-sdk",
    "databricks-sql-connector",
    "pandas",
    "ipython",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/db_magic"]
